% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClonalityPackage.R
\name{addMutIDtoTableAndGetMappingMutIDtoMut}
\alias{addMutIDtoTableAndGetMappingMutIDtoMut}
\title{Adds mutation IDs to the table of mutations using only identityColumns}
\usage{
addMutIDtoTableAndGetMappingMutIDtoMut(eventDataFrame, identityColumns)
}
\arguments{
\item{eventDataFrame}{Table of mutations in the dataframe format}

\item{identityColumns}{The columns that identify a event. All events with the same values in the
identyColumns will be grouped together.}
}
\description{
Adds mutation IDs to the table of mutations using only identityColumns
}
