% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClonalityPackage.R
\name{clusterCloseMutations}
\alias{clusterCloseMutations}
\title{Cluster mutations within ditanceTolerance}
\usage{
clusterCloseMutations(eventDataFrame, identityColumns, ditanceTolerance)
}
\arguments{
\item{eventDataFrame}{Table of mutations in the dataframe format}

\item{identityColumns}{The columns that identify a event. All events with the same values in the
identyColumns will be grouped together.}

\item{ditanceTolerance}{Distance tolerance}
}
\description{
Cluster mutations within ditanceTolerance
}
