% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClonalityPackage.R
\name{computeSharedMutations}
\alias{computeSharedMutations}
\title{Compute table of shared mutations}
\usage{
computeSharedMutations(pairs, eventDataFrame, identityColumns)
}
\arguments{
\item{pairs}{A dataframe of sample pairs with sample identifiers in columns sample1 and sample2.}

\item{eventMatrix}{A binary matrix indicating events (columns) in samples (rows).}

\item{eventExpection}{A matrix indicating the expectation of events (columns) in samples (rows).}

\item{scorefun}{The function used to score a pair of samples.}
}
\value{
pairs with an added column scores containing the clonality scores.
}
\description{
Compute table of shared mutations
}
