% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClonalityPackage.R
\name{eventDataFrameToMatrix}
\alias{eventDataFrameToMatrix}
\title{Converts a dataframe of events to a matrix of events by samples.}
\usage{
eventDataFrameToMatrix(eventDataFrame, identityColumns, sampleColumn = NULL)
}
\arguments{
\item{eventDataFrame}{Table of mutations in the dataframe format}

\item{identityColumns}{The columns that identify a event. All events with the same values in the
identyColumns will be grouped together.}

\item{sampleColumn}{The column that contains the sample identifiers.}
}
\description{
Converts a dataframe of events to a matrix of events by samples.
}
