% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClonalityPackage.R
\name{scoreClonality}
\alias{scoreClonality}
\title{Score the clonality of a single pair of samples.}
\usage{
scoreClonality(events1, events2, eventExpectation1, eventExpectation2)
}
\arguments{
\item{events1}{A binary vector with the events of the first sample.}

\item{events2}{A binary vector with the events of the second sample.}

\item{eventExpectation1}{A binary vector with the expectation of all events of the first sample, row of the eventExpectation for the first sample.}

\item{eventExpectation2}{A binary vector with the expectation of all events of the second sample, row of the eventExpectation for the second sampke.}
}
\description{
Score the clonality of a single pair of samples.
}
