% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClonalityPackage.R
\name{testClonalityScores}
\alias{testClonalityScores}
\title{Generate a reference distribution by computing the score for all random pairs from different
patients and calculate p-value.}
\usage{
testClonalityScores(
  pairs,
  eventMatrix,
  eventExpectation,
  scoreFun = scoreClonalityn
)
}
\arguments{
\item{pairs}{A dataframe of sample pairs with sample identifiers in columns sample1 and sample2.
Column Patient is optional. If it is not available it is assumed all pairs are from different
patients. Column score gives the score for the pairs.}

\item{events}{A matrix indicating the expectation of events (columns) in samples (rows).}

\item{scorefun}{The function used to score a pair of samples.}
}
\value{
pairs with an added column p containing p-values from the randomization test.
}
\description{
Generate a reference distribution by computing the score for all random pairs from different
patients and calculate p-value.
}
